refactor: extract catalog surface policy
(cherry picked from commit 164826b514de22609e2e08b4e7d3a5e51a84023f)
